.pyenv/requirements.txt
alabaster
argh
astroid
atomicwrites
attrs
autopep8
Babel
backcall
bleach
certifi
cffi
chardet
cloudpickle
cryptography
cycler
decorator
defusedxml
diff-match-patch
docutils
entrypoints
flake8
future
helpdev
idna
imagesize
importlib-metadata
intervaltree
ipykernel
ipython
ipython-genutils
isort
japanize-matplotlib
jedi
jeepney
Jinja2
jsonschema
jupyter-client
jupyter-core
keyring
kiwisolver
lazy-object-proxy
MarkupSafe
matplotlib
mccabe
mistune
more-itertools
nbconvert
nbformat
numpy
numpydoc
packaging
pandas
pandocfilters
parso
pathtools
pexpect
pickleshare
pkg-resources
pluggy
prompt-toolkit
psutil
ptyprocess
pycodestyle
pycparser
pydocstyle
pyflakes
Pygments
pylint
pyparsing
PyQt5
PyQt5-sip
PyQtWebEngine
pyrsistent
PySoundFile
python-dateutil
python-jsonrpc-server
python-language-server
pytz
pyxdg
PyYAML
pyzmq
QDarkStyle
QtAwesome
qtconsole
QtPy
requests
rope
scipy
seaborn
SecretStorage
six
snowballstemmer
sortedcontainers
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
testpath
tornado
traitlets
typed-ast
ujson
urllib3
watchdog
wcwidth
webencodings
wrapt
wurlitzer
yapf
zipp
